Add files via upload
This project, created during my 3rd year, is a dynamic showcase of my photography skills, built using HTML, CSS, and JavaScript. The website features a user-friendly interface with smooth navigation and interactive elements like image sliders and lightboxes. Each category, such as Nature, Portraits, and Street Photography, is thoughtfully curated to highlight my versatility. The responsive design ensures it looks great on all devices, while the minimalist layout keeps the focus on the visuals. It’s a perfect blend of creativity and technical expertise, reflecting my growth as both a photographer and a developer.